js file: union methods in 1 file
