+++ piskel-clone-app/src/app/components/frame-container/frame-method.js
@@ -1,6 +1,7 @@
 import {addActiveClassToFrame} from "../../utils/frameUpdate";
 import createFrame from '../createFrame';
 import {pasteImage} from "../../utils/drawImage";
+import {updateFramesState} from "../../utils/frameUpdate";
 
 let frameContainers; 
 let canvas;
@@ -19,4 +20,20 @@ function addFrame(frames) {
     frames.push(frame);
 }
 
-export default addFrame;
+function dublicateFrame(copiedFrame, canvas) {
+    const frameNum = copiedFrame.children[0].textContent;
+    const frameContainer = createFrame(Number(frameNum) + 1);
+    const frame = frameContainer.firstElementChild;
+    const copiedFrameContainer = copiedFrame.parentElement;
+    copiedFrameContainer.after(frameContainer);
+
+    const url = copiedFrame.children[1].toDataURL();
+
+    pasteImage(url, frame.children[1]);
+    pasteImage(url, canvas);
+
+    return updateFramesState();
+}
+
+
+export {dublicateFrame, addFrame};
