--- piskel-clone-app/src/app/entities/helpers/getImageFromFrames.js
@@ -1,9 +1,0 @@
-function getImageFromFrames() {
-    const frameContainers = document.querySelectorAll('.frame-container');
-    const imageFromFrames = Array.from(frameContainers).map(frame => 
-        frame.firstElementChild.children[1].toDataURL());
-
-    return imageFromFrames;
-}
-
-export default getImageFromFrames;
